fix(csa-session): add fallback_chain to SessionResult; use atomic write path
Finding 1: `write_fallback_chain_to_result_toml` used `std::fs::write`
to overwrite an already-published result.toml non-atomically, risking
corruption under concurrent readers (`csa session wait/result/list`).
Fixed by routing through `load_result` + `save_result`, which uses
the existing `write_file_atomically` helper internally.

Finding 2: `fallback_chain` was injected as raw TOML outside the
`SessionResult` struct, so it was invisible to `csa session result --json`
and the typed API. Added `fallback_chain: Option<Vec<FallbackAttempt>>`
with `#[serde(default, skip_serializing_if = "Option::is_none")]`
so the field is round-tripped correctly and exposed via the structured API.

Also splits `manager_tests_tail.rs` (797→806 lines after the struct
field addition) to stay within the 800-line monolith limit.
