+++ crates/cli-sub-agent/src/pipeline_tests_post_exec_audit.rs
@@ -94,6 +94,7 @@ fn apply_repo_write_audit_to_result_populates_manager_sidecar_sections() {
         events_count: 0,
         artifacts: vec![],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
     let audit = RepoWriteAudit {
@@ -404,6 +405,7 @@ fn audit_failure_does_not_fail_execution() {
         events_count: 0,
         artifacts: vec![],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
 
@@ -525,6 +527,7 @@ fn reused_session_audit_uses_per_execution_baseline_not_session_creation() {
         events_count: 0,
         artifacts: vec![],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
 
@@ -625,6 +628,7 @@ fn first_execution_falls_back_to_session_creation_baseline_when_per_exec_capture
         events_count: 0,
         artifacts: vec![],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
