+++ crates/cli-sub-agent/src/preflight_state_dir.rs
@@ -707,6 +707,7 @@ mod tests {
                 events_count: 0,
                 artifacts: vec![SessionArtifact::new("output/summary.md")],
                 peak_memory_mb: None,
+                fallback_chain: None,
                 manager_fields: Default::default(),
             },
         )
