+++ crates/cli-sub-agent/src/session_cmds_tests.rs
@@ -56,6 +56,7 @@ fn make_result(status: &str, exit_code: i32) -> SessionResult {
         events_count: 0,
         artifacts: Vec::new(),
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     }
 }
