+++ crates/csa-session/src/manager_tests_audit.rs
@@ -15,6 +15,7 @@ fn test_save_result_persists_manager_fields_sidecar() {
         events_count: 0,
         artifacts: vec![],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: crate::result::SessionManagerFields {
             artifacts: Some(
                 toml::toml! {
