+++ crates/csa-session/src/manager_tests_result_view.rs
@@ -64,6 +64,7 @@ fn test_load_result_view_ignores_orphaned_manager_sidecar() {
         events_count: 1,
         artifacts: vec![crate::result::SessionArtifact::new("output/acp-events.jsonl")],
         peak_memory_mb: None,
+            fallback_chain: None,
             manager_fields: Default::default(),
     };
     save_result_in(
@@ -120,6 +121,7 @@ fn test_load_result_merges_manager_sidecar_sections_into_runtime_result() {
         events_count: 1,
         artifacts: vec![crate::result::SessionArtifact::new("output/acp-events.jsonl")],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
     std::fs::write(
@@ -238,6 +240,7 @@ fn test_manager_sidecar_roundtrip_preserves_full_sa_schema() {
         events_count: 1,
         artifacts: vec![crate::result::SessionArtifact::new("output/acp-events.jsonl")],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
     let input_sidecar = toml::Value::Table(toml::toml! {
@@ -387,6 +390,7 @@ fn test_load_result_without_sidecar_keeps_manager_fields_empty() {
         events_count: 1,
         artifacts: vec![crate::result::SessionArtifact::new("output/acp-events.jsonl")],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
     save_result_in(
@@ -423,6 +427,7 @@ fn test_save_result_with_empty_manager_fields_preserves_existing_sidecar() {
         events_count: 1,
         artifacts: vec![crate::result::SessionArtifact::new("output/acp-events.jsonl")],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: crate::result::SessionManagerFields {
             report: Some(
                 toml::toml! {
@@ -448,6 +453,7 @@ fn test_save_result_with_empty_manager_fields_preserves_existing_sidecar() {
     );
 
     let clean_result = crate::result::SessionResult {
+        fallback_chain: None,
         manager_fields: Default::default(),
         ..populated_result.clone()
     };
@@ -494,6 +500,7 @@ fn test_clear_manager_sidecar_removes_existing_sidecar() {
         events_count: 1,
         artifacts: vec![crate::result::SessionArtifact::new("output/acp-events.jsonl")],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: crate::result::SessionManagerFields {
             report: Some(
                 toml::toml! {
@@ -545,6 +552,7 @@ fn test_load_result_with_malformed_manager_sidecar_is_non_fatal() {
         events_count: 1,
         artifacts: vec![crate::result::SessionArtifact::new("output/acp-events.jsonl")],
         peak_memory_mb: None,
+        fallback_chain: None,
         manager_fields: Default::default(),
     };
     std::fs::write(
