+++ crates/csa-session/src/manager_tests_tail_2.rs
@@ -1,0 +1,349 @@
+
+#[test]
+fn test_list_artifacts() {
+    let td = tempdir().unwrap();
+    let state = create_session_in(td.path(), td.path(), None, None, Some("codex")).unwrap();
+    let dir = get_session_dir_in(td.path(), &state.meta_session_id);
+    std::fs::write(dir.join("output/report.txt"), "test").unwrap();
+    std::fs::write(dir.join("output/diff.patch"), "test").unwrap();
+    std::fs::write(dir.join("output/acp-events.jsonl"), "{\"ts\":\"2026-01-01T00:00:00Z\"}\n")
+        .unwrap();
+    let artifacts = list_artifacts_in(td.path(), &state.meta_session_id).unwrap();
+    assert_eq!(artifacts.len(), 3);
+    assert!(artifacts.contains(&"acp-events.jsonl".to_string()));
+    assert!(artifacts.contains(&"diff.patch".to_string()));
+    assert!(artifacts.contains(&"report.txt".to_string()));
+}
+
+#[test]
+fn test_status_from_exit_code() {
+    use crate::result::SessionResult;
+    assert_eq!(SessionResult::status_from_exit_code(0), "success");
+    assert_eq!(SessionResult::status_from_exit_code(1), "failure");
+    assert_eq!(SessionResult::status_from_exit_code(137), "signal");
+    assert_eq!(SessionResult::status_from_exit_code(143), "signal");
+}
+
+#[test]
+fn test_save_session_in_explicit_base() {
+    let td = tempdir().unwrap();
+    let mut state =
+        create_session_in(td.path(), td.path(), Some("Explicit save"), None, None).unwrap();
+    state.description = Some("Modified".to_string());
+    save_session_in(td.path(), &state).unwrap();
+    let loaded = load_session_in(td.path(), &state.meta_session_id).unwrap();
+    assert_eq!(loaded.description, Some("Modified".to_string()));
+}
+
+#[test]
+fn test_list_sessions_empty_and_missing() {
+    let td = tempdir().unwrap();
+    assert!(list_all_sessions_in(td.path()).unwrap().is_empty());
+    assert!(list_sessions_in(td.path(), None).unwrap().is_empty());
+}
+
+#[test]
+fn test_delete_nonexistent_session() {
+    let td = tempdir().unwrap();
+    std::fs::create_dir_all(td.path().join("sessions")).unwrap();
+    let r = delete_session_in(td.path(), &crate::validate::new_session_id());
+    assert!(r.unwrap_err().to_string().contains("not found"));
+}
+
+#[test]
+fn test_load_nonexistent_session() {
+    let td = tempdir().unwrap();
+    let r = load_session_in(td.path(), &crate::validate::new_session_id());
+    assert!(r.unwrap_err().to_string().contains("not found"));
+}
+
+#[test]
+fn test_update_last_accessed_advances_timestamp() {
+    let td = tempdir().unwrap();
+    let state = create_session_in(td.path(), td.path(), Some("ts"), None, None).unwrap();
+    let t0 = state.last_accessed;
+    std::thread::sleep(std::time::Duration::from_millis(10));
+    let mut s = load_session_in(td.path(), &state.meta_session_id).unwrap();
+    s.last_accessed = Utc::now();
+    save_session_in(td.path(), &s).unwrap();
+    let s2 = load_session_in(td.path(), &state.meta_session_id).unwrap();
+    assert!(s2.last_accessed > t0);
+}
+
+#[test]
+fn test_list_artifacts_empty_output() {
+    let td = tempdir().unwrap();
+    let state = create_session_in(td.path(), td.path(), None, None, None).unwrap();
+    assert!(
+        list_artifacts_in(td.path(), &state.meta_session_id)
+            .unwrap()
+            .is_empty()
+    );
+}
+
+#[test]
+fn test_operations_with_invalid_session_id() {
+    let td = tempdir().unwrap();
+    let bad = "not-a-valid-ulid";
+    assert!(load_session_in(td.path(), bad).is_err());
+    assert!(delete_session_in(td.path(), bad).is_err());
+    assert!(load_metadata_in(td.path(), bad).is_err());
+    assert!(validate_tool_access_in(td.path(), bad, "codex").is_err());
+}
+
+const ETXTBSY_RAW_OS_ERROR: i32 = 26;
+
+fn git_command_retries_exhausted(args: &[&str], attempts: usize, err: &std::io::Error) -> ! {
+    panic!(
+        "git {args:?} failed after {attempts} attempts: {err}"
+    );
+}
+
+fn is_text_file_busy(error: &std::io::Error) -> bool {
+    error.raw_os_error() == Some(ETXTBSY_RAW_OS_ERROR)
+}
+
+fn etxtbsy_backoff(attempt: usize) -> std::time::Duration {
+    let delay_ms = 25_u64.saturating_mul(1_u64 << attempt.min(3));
+    std::time::Duration::from_millis(delay_ms)
+}
+
+fn run_git_output(dir: &std::path::Path, args: &[&str]) -> std::process::Output {
+    let max_attempts = 4;
+    for attempt in 0..max_attempts {
+        match std::process::Command::new("git")
+            .args(args)
+            .current_dir(dir)
+            .output()
+        {
+            Ok(output) => return output,
+            Err(err) if is_text_file_busy(&err) && attempt + 1 < max_attempts => {
+                std::thread::sleep(etxtbsy_backoff(attempt));
+            }
+            Err(err) => git_command_retries_exhausted(args, attempt + 1, &err),
+        }
+    }
+    unreachable!("retry loop should have returned or panicked")
+}
+
+fn run_git(dir: &std::path::Path, args: &[&str]) {
+    let output = run_git_output(dir, args);
+    assert!(output.status.success(), "git {args:?} failed");
+}
+
+#[test]
+fn test_create_session_records_branch_in_git_repo() {
+    let td = tempdir().unwrap();
+    run_git(td.path(), &["init"]);
+    run_git(td.path(), &["config", "user.email", "test@example.com"]);
+    run_git(td.path(), &["config", "user.name", "Test User"]);
+    std::fs::write(td.path().join("README.md"), "test").unwrap();
+    run_git(td.path(), &["add", "README.md"]);
+    run_git(td.path(), &["commit", "-m", "init"]);
+    run_git(td.path(), &["checkout", "-b", "feat/session-discovery"]);
+    let state = create_session_in(td.path(), td.path(), Some("git"), None, None).unwrap();
+    assert_eq!(state.branch.as_deref(), Some("feat/session-discovery"));
+}
+
+#[test]
+fn test_find_sessions_multi_condition_filtering() {
+    let td = tempdir().unwrap();
+    let _xdg = ScopedXdgOverride::new(&td);
+    let mut s1 = create_session_in(td.path(), td.path(), Some("S1"), None, None).unwrap();
+    s1.branch = Some("feature/a".to_string());
+    s1.phase = SessionPhase::Available;
+    s1.task_context.task_type = Some("plan".to_string());
+    s1.last_accessed = Utc::now() - chrono::Duration::minutes(5);
+    s1.tools.insert(
+        "codex".to_string(),
+        crate::state::ToolState {
+            provider_session_id: None,
+            last_action_summary: "Plan".to_string(),
+            last_exit_code: 0,
+            updated_at: Utc::now(),
+            tool_version: None,
+            token_usage: None,
+        },
+    );
+    save_session_in(td.path(), &s1).unwrap();
+
+    let mut s2 = create_session_in(td.path(), td.path(), Some("S2"), None, None).unwrap();
+    s2.branch = Some("feature/a".to_string());
+    s2.phase = SessionPhase::Available;
+    s2.task_context.task_type = Some("review".to_string());
+    s2.last_accessed = Utc::now();
+    s2.tools.insert(
+        "codex".to_string(),
+        crate::state::ToolState {
+            provider_session_id: None,
+            last_action_summary: "Review".to_string(),
+            last_exit_code: 0,
+            updated_at: Utc::now(),
+            tool_version: None,
+            token_usage: None,
+        },
+    );
+    save_session_in(td.path(), &s2).unwrap();
+
+    let mut s3 = create_session_in(td.path(), td.path(), Some("S3"), None, None).unwrap();
+    s3.branch = Some("feature/b".to_string());
+    s3.phase = SessionPhase::Available;
+    s3.task_context.task_type = Some("plan".to_string());
+    s3.last_accessed = Utc::now() - chrono::Duration::minutes(1);
+    s3.tools.insert(
+        "gemini-cli".to_string(),
+        crate::state::ToolState {
+            provider_session_id: None,
+            last_action_summary: "Plan".to_string(),
+            last_exit_code: 0,
+            updated_at: Utc::now(),
+            tool_version: None,
+            token_usage: None,
+        },
+    );
+    save_session_in(td.path(), &s3).unwrap();
+
+    let sessions = find_sessions_in(
+        td.path(),
+        Some(td.path()),
+        Some("feature/a"),
+        Some("plan"),
+        Some(SessionPhase::Available),
+        Some(&["codex"]),
+    )
+    .unwrap();
+
+    assert_eq!(sessions.len(), 1);
+    assert_eq!(sessions[0].meta_session_id, s1.meta_session_id);
+}
+
+#[test]
+fn test_find_sessions_sorts_desc_and_limits_to_ten() {
+    let td = tempdir().unwrap();
+    let _xdg = ScopedXdgOverride::new(&td);
+    for i in 0..12 {
+        let mut session = create_session_in(td.path(), td.path(), Some("S"), None, None).unwrap();
+        session.branch = Some("feature/a".to_string());
+        session.task_context.task_type = Some("plan".to_string());
+        session.phase = SessionPhase::Available;
+        session.last_accessed = Utc::now() - chrono::Duration::minutes(i);
+        save_session_in(td.path(), &session).unwrap();
+    }
+
+    let sessions = find_sessions_in(
+        td.path(),
+        Some(td.path()),
+        Some("feature/a"),
+        Some("plan"),
+        Some(SessionPhase::Available),
+        None,
+    )
+    .unwrap();
+
+    assert_eq!(sessions.len(), 10);
+    assert!(
+        sessions
+            .windows(2)
+            .all(|pair| pair[0].last_accessed >= pair[1].last_accessed)
+    );
+}
+
+#[test]
+fn test_global_exact_finds_cross_project_session() {
+    let td = tempdir().unwrap();
+    let _xdg = ScopedXdgOverride::new(&td);
+
+    // Create two "project" directories under the temp dir
+    let project_a_root = td.path().join("project_a");
+    let project_b_root = td.path().join("project_b");
+
+    // Create a session under project_a's session root
+    let session_a =
+        create_session_in(&project_a_root, &project_a_root, Some("from project A"), None, None)
+            .unwrap();
+
+    // Create a session under project_b's session root
+    let session_b =
+        create_session_in(&project_b_root, &project_b_root, Some("from project B"), None, None)
+            .unwrap();
+
+    // Verify we can load each session from its own root
+    let loaded_a = load_session_in(&project_a_root, &session_a.meta_session_id).unwrap();
+    assert_eq!(loaded_a.description, Some("from project A".to_string()));
+
+    let loaded_b = load_session_in(&project_b_root, &session_b.meta_session_id).unwrap();
+    assert_eq!(loaded_b.description, Some("from project B".to_string()));
+
+    // Verify we CANNOT load project_a's session from project_b's root (cross-project)
+    let cross_load = load_session_in(&project_b_root, &session_a.meta_session_id);
+    assert!(cross_load.is_err());
+}
+
+#[test]
+fn test_extract_project_path_from_state_content() {
+    let content = r#"
+meta_session_id = "01HTESTABCDEFGHIJKLMNOPQR"
+project_path = "/home/user/project-a"
+description = "test"
+"#;
+    // Test the extract function indirectly via a state.toml-like content
+    assert!(content.contains("project_path"));
+    for line in content.lines() {
+        let trimmed = line.trim();
+        if let Some(rest) = trimmed.strip_prefix("project_path") {
+            let rest = rest.trim();
+            if let Some(rest) = rest.strip_prefix('=') {
+                let value = rest.trim().trim_matches('"');
+                assert_eq!(value, "/home/user/project-a");
+            }
+        }
+    }
+}
+
+#[test]
+fn test_list_all_sessions_from_multiple_roots() {
+    let td = tempdir().unwrap();
+    let _xdg = ScopedXdgOverride::new(&td);
+    let root_a = td.path().join("root_a");
+    let root_b = td.path().join("root_b");
+
+    let s1 = create_session_in(&root_a, &root_a, Some("session A1"), None, None).unwrap();
+    let s2 = create_session_in(&root_a, &root_a, Some("session A2"), None, None).unwrap();
+    let s3 = create_session_in(&root_b, &root_b, Some("session B1"), None, None).unwrap();
+
+    // Each root should have its own sessions
+    let sessions_a = list_all_sessions_in_readonly(&root_a).unwrap();
+    assert_eq!(sessions_a.len(), 2);
+
+    let sessions_b = list_all_sessions_in_readonly(&root_b).unwrap();
+    assert_eq!(sessions_b.len(), 1);
+
+    // Verify session IDs
+    let ids_a: Vec<&str> = sessions_a.iter().map(|s| s.meta_session_id.as_str()).collect();
+    assert!(ids_a.contains(&s1.meta_session_id.as_str()));
+    assert!(ids_a.contains(&s2.meta_session_id.as_str()));
+    assert_eq!(sessions_b[0].meta_session_id, s3.meta_session_id);
+}
+
+#[test]
+fn test_prefix_stays_project_scoped() {
+    let td = tempdir().unwrap();
+    let _xdg = ScopedXdgOverride::new(&td);
+    let root_a = td.path().join("root_a");
+    let root_b = td.path().join("root_b");
+
+    let s1 = create_session_in(&root_a, &root_a, Some("A"), None, None).unwrap();
+    let _s2 = create_session_in(&root_b, &root_b, Some("B"), None, None).unwrap();
+
+    // Full-ID resolution should only find sessions in the specified root
+    let sessions_dir_a = root_a.join("sessions");
+    let resolved =
+        crate::validate::resolve_session_prefix(&sessions_dir_a, &s1.meta_session_id).unwrap();
+    assert_eq!(resolved, s1.meta_session_id);
+
+    // Same full ID should NOT match in root_b's sessions dir
+    let sessions_dir_b = root_b.join("sessions");
+    let result = crate::validate::resolve_session_prefix(&sessions_dir_b, &s1.meta_session_id);
+    assert!(result.is_err());
+}
+
